+++ .drone/scripts/run-unit-tests.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/bash
 set -e
 
-if git log -1 --pretty='%s' | grep '\[TEST SKIP\]$'; then
+if (git log -1 --pretty='%s' | grep '\[TEST SKIP\]$') || ([[ "${TEST_SKIP:+x}" == 'x' ]]); then
     echo "Skipping unit tests..."
     exit 0
 fi